Complete playlist duration utility for playlist duration component
