+++ client/src/components/playlist-duration/playlist-duration.component.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
-import getSeconds from './playlist-duration.util';
+import getTime from './playlist-duration.util';
 import './playlist-duration.styles.css';
 
 const PlaylistDuration = () => {
@@ -8,25 +8,7 @@ const PlaylistDuration = () => {
     ({ likedSongs: { likedSongs } }) => likedSongs,
   );
 
-  const totalSeconds = getSeconds(likedSongsList);
-
-  let totalMinutes = likedSongsList.reduce(
-    (accum, song) => accum + song.minutes,
-    0,
-  );
-  const extraMinutes = Math.floor(totalSeconds / 60);
-  totalMinutes += extraMinutes;
-
-  let remainingMinutes = 0;
-  let totalHours = 0;
-
-  if (totalMinutes >= 60) {
-    totalHours += Math.floor(totalMinutes / 60);
-    remainingMinutes = totalMinutes % 60;
-    totalMinutes = remainingMinutes;
-  }
-
-  const remainingSeconds = totalSeconds % 60;
+  const [totalHours, totalMinutes, remainingSeconds] = getTime(likedSongsList);
 
   return (
     <div className="playlist-duration-container">
